Fix: unwanted layer animation caused by CA
Whenever updating geometoric parameters of CALayer, it animates automatically using the default Core Animation settings. That's why the video player (=renderingLayer) animates without any UIKit animation blocks. So instead, I swapped the base layer of the VersaPlayerView, which is managed by NSView / UIView so that this will not automatically animates.